requirements.txt
# Data fetching
yfinance>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
mplfinance>=0.12.0
pillow>=10.0.0

# Machine learning
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# NLP / Embeddings
transformers>=4.30.0
accelerate>=0.20.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.0

# Optional
pandas-ta>=0.3.0
requests>=2.31.0
